backend/requirements.txt
Django
django-cleanup
django-cors-headers
django-filter
djangorestframework
django-phonenumber-field[phonenumbers]
Pillow
pyinstaller
PyJWT
cherrypy
